feat: home 페이지 구조 변경
